pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "podtext"
version = "0.1.0"
description = "CLI tool for podcast discovery, transcription, and AI-powered analysis"
readme = "README.md"
requires-python = ">=3.13"
license = "MIT"
authors = [
    { name = "podtext contributors" }
]
keywords = ["podcast", "transcription", "whisper", "claude", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]

dependencies = [
    "click>=8.1.0",
    "feedparser>=6.0.0",
    "httpx>=0.27.0",
    "anthropic>=0.40.0",
    "mlx-whisper>=0.4.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.5.0",
    "respx>=0.21.0",
]

[project.scripts]
podtext = "podtext.cli.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["podtext"]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.coverage.run]
source = ["podtext"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
]
